refactor : Exception 예외 핸들러에서 응답해주는 메세지를 '서버에서 오류가 발생했습니다.' 로 수정
